backend/requirements.txt
# Core Framework
flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21
alembic==1.12.0

# Image Processing
Pillow==10.0.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.4.2

# HTTP Client
requests>=2.31.0

# File Handling
python-multipart==0.0.6

# Server
gunicorn==21.2.0

# Testing & Development
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-env==1.0.1
black==23.9.1
mypy==1.5.1
isort==5.12.0
flake8==6.1.0
pylint==3.0.0

# Remove until implemented:
# sentry-sdk==1.31.0
# structlog==23.1.0
# python-jose==3.3.0
# bcrypt==4.0.1 

# Add these to existing requirements
python-jose[cryptography]>=3.3.0  # For JWT handling
passlib[bcrypt]>=1.7.4  # For password hashing
python-multipart>=0.0.5  # For form data processing
google-auth>=2.22.0  # For Google OAuth
PyJWT==2.8.0